Workspaces Server

+++ deployment/postgres-init-db/sql/views/migrations/applied_migrations_view.sql
@@ -1,3 +1,3 @@
 create view migrations.applied_migrations as
-  select migration_id::int
+  select migration_id
   from migrations.schema_migrations;
